+++ EliteJournalReader/Events/CrewFireEvent.cs
@@ -9,7 +9,7 @@ namespace EliteJournalReader.Events
 {
     //When Written: when dismissing a member of crew
     //Parameters:
-    //•	Name
+    //ï¿½	Name
     public class CrewFireEvent : JournalEvent<CrewFireEvent.CrewFireEventArgs>
     {
         public CrewFireEvent() : base("CrewFire") { }
@@ -17,7 +17,7 @@ namespace EliteJournalReader.Events
         public class CrewFireEventArgs : JournalEventArgs
         {
             public string Name { get; set; }
-            public int CrewID { get; set; }
+            public long CrewID { get; set; }
         }
     }
 }
