+++ EliteJournalReader/Events/CrewHireEvent.cs
@@ -9,14 +9,14 @@ namespace EliteJournalReader.Events
 {
     //When Written: when engaging a new member of crew
     //Parameters:
-    //•	Name
+    //ï¿½	Name
     public class CrewHireEvent : JournalEvent<CrewHireEvent.CrewHireEventArgs>
     {
         public CrewHireEvent() : base("CrewHire") { }
 
         public class CrewHireEventArgs : JournalEventArgs
         {
-            public int CrewID { get; set; }
+            public long CrewID { get; set; }
             public string Name { get; set; }
             public string Faction { get; set; }
             public int Cost { get; set; }
