+++ EliteJournalReader/Events/CrewLaunchFighterEvent.cs
@@ -9,7 +9,7 @@ namespace EliteJournalReader.Events
 {
     //When written: when in multicrew, in Helm player's log, when a crew member launches a fighter
     //Parameters:
-    //•	Crew: name of crew member launching in fighter
+    //ï¿½	Crew: name of crew member launching in fighter
     public class CrewLaunchFighterEvent : JournalEvent<CrewLaunchFighterEvent.CrewLaunchFighterEventArgs>
     {
         public CrewLaunchFighterEvent() : base("CrewLaunchFighter") { }
@@ -17,8 +17,8 @@ namespace EliteJournalReader.Events
         public class CrewLaunchFighterEventArgs : JournalEventArgs
         {
             public string Crew { get; set; }
-            public int CrewID { get; set; }
-            public int ID { get; set; }
+            public long CrewID { get; set; }
+            public long ID { get; set; }
         }
     }
 }
