+++ EliteJournalReader/Events/CrewMemberJoinsEvent.cs
@@ -9,7 +9,7 @@ namespace EliteJournalReader.Events
 {
     //When written: When another player joins your ship's crew
     //Parameters:
-    //•	Crew: player's commander name
+    //ï¿½	Crew: player's commander name
     public class CrewMemberJoinsEvent : JournalEvent<CrewMemberJoinsEvent.CrewMemberJoinsEventArgs>
     {
         public CrewMemberJoinsEvent() : base("CrewMemberJoins") { }
@@ -17,7 +17,7 @@ namespace EliteJournalReader.Events
         public class CrewMemberJoinsEventArgs : JournalEventArgs
         {
             public string Crew { get; set; }
-            public int CrewID { get; set; }
+            public long CrewID { get; set; }
         }
     }
 }
