+++ EliteJournalReader/Events/CrewMemberRoleChangeEvent.cs
@@ -9,8 +9,8 @@ namespace EliteJournalReader.Events
 {
     //When written: When another player joins your ship's crew
     //Parameters:
-    //•	Crew: player's commander name
-    //•	Role: selected role
+    //ï¿½	Crew: player's commander name
+    //ï¿½	Role: selected role
     public class CrewMemberRoleChangeEvent : JournalEvent<CrewMemberRoleChangeEvent.CrewMemberRoleChangeEventArgs>
     {
         public CrewMemberRoleChangeEvent() : base("CrewMemberRoleChange") { }
@@ -18,7 +18,7 @@ namespace EliteJournalReader.Events
         public class CrewMemberRoleChangeEventArgs : JournalEventArgs
         {
             public string Crew { get; set; }
-            public int CrewID { get; set; }
+            public long CrewID { get; set; }
             public string Role { get; set; }
         }
     }
