+++ EliteJournalReader/Events/ScreenshotEvent.cs
@@ -9,13 +9,13 @@ namespace EliteJournalReader.Events
 {
     //When Written: when a screen snapshot is saved
     //Parameters: 
-    //•	Filename: filename of screenshot
-    //•	Width: size in pixels
-    //•	Height: size in pixels
-    //•	System: current star system
-    //•	Body: name of nearest body
-    //•	Latitude: The latitude and longitude will be included if on a planet or in low-altitude flight
-    //•	Longitude
+    //ï¿½	Filename: filename of screenshot
+    //ï¿½	Width: size in pixels
+    //ï¿½	Height: size in pixels
+    //ï¿½	System: current star system
+    //ï¿½	Body: name of nearest body
+    //ï¿½	Latitude: The latitude and longitude will be included if on a planet or in low-altitude flight
+    //ï¿½	Longitude
     public class ScreenshotEvent : JournalEvent<ScreenshotEvent.ScreenshotEventArgs>
     {
         public ScreenshotEvent() : base("Screenshot") { }
@@ -23,7 +23,7 @@ namespace EliteJournalReader.Events
         public class ScreenshotEventArgs : JournalEventArgs
         {
             public string Filename { get; set; }
-            public int Width { get; set; }
+            public long Width { get; set; }
             public int Height { get; set; }
             public string System { get; set; }
             public string Body { get; set; }
